json: do not marshal callsites root children
Nodes which are children of the root node do not have a meaningful
callsite. Generally, the callsite is variable from one process to the
next, and thus could not possibly be listed in the dwarf table.

So far the only program constructs that lead to children of root are
so-called top-level functions, which include main and functions
launched with pthread_create. There may be others, but in any case
they are not expected to have callsites in dwarf.