server: log errors
This required server.c to depend on socket.h, which contains the logprint
function. This had a chain of consequences for code organization.

In particular, it made socket.h's dependence on node.h (by way of
functions sendstacktrace and sendprofile, which refer to Node) seem
inappropriate, so the misplaced functions were moved to node.c. This
required node.c to depend on buf.h.

A dependence by node.c on json.h was avoided by using function pointer
parameters. The intent is not to abstract out the encoding, as might
be evident by the JSON string literals in these functions. It's a
convenient hack to avoid more header inclusions.