buf: internalize error handling
Buf now contains an error member that is nonzero if an error has
occurred. This allows functions that work with Bufs to examine
the current error state and exit immediately, so that their
callers do not have to constantly check for errors.

The effect on readability is significant. It also reduces the
proportion of error checking code, so that our coverage
numbers will probably increase.

json:

Calls to write have been moved out of the send* functions, and
these functions have been renamed marshal*. Thus, json functions
only marshal into Bufs, and it is the caller's choice what to
do with the data.