node: unexport newNode
newNode should not be used by other modules. Unexporting it prevents
them from misusing it. This then requires them to construct Node
trees in the supported way: by using push and pop.

This requires the json tests to be rewritten, as they used
newNode. Each test case now consists of a stack operator, as it's
not possible to construct trees declaratively.