Add thread test; fix several bugs
thread_test:

is a multi-threaded program to discover races when running under
Valgrind. The test requires a patched version of the client,
otherwise the emission request logic will not get covered.

server:

pthread_kill appears not to be the most appropriate way to force
the server thread to exit. pthread_cancel is better, as it uses
synchronous cancellation points built into libc, which includes
read(2).

json, node:

json.c was accessing node fields without a lock. Some of these
fields are native integers and thus probably provide atomic access
on most platforms, but we add lock calls to appease Helgrind and DRD.

Nevertheless, our use of locks is still subject to the TOCTTOU
bug[1], as we're not holding the locks long enough.

[1]: https://en.wikipedia.org/wiki/Time_of_check_to_time_of_use

json:

Running thread_test under a patched exposed a serialization bug.